+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Version 0.1.10 (February 24th 2019)
+-----------------------------
+ * Added base64 encoding option
+ * Added ability to set headers into HTTP requests methods
+ 
 Version 0.1.9 (December 24th 2018)
 -----------------------------
  * Library is lighter